Support Vector Maschine
Added the support vector machine code
It is currently working with manually setting params. This will have to be
adapted to all data and then run in a loop to run over night I believe. It
is however quite fast for myData to run. Takes like 5 seconds. So most
likely AllData won't be that long either.

Haven't Tried to apply PCA yet.